+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 3.0.0 (unreleased)
 
+* `versions` – Remove deprecated `:version_names`, `Shrine.version_names` and `Shrine.version?` (@janko)
+
 * `validation_helpers` – Remove support for regexes in MIME type or extension validations (@janko)
 
 * `validation_helpers` – Don't require `#width` and `#height` methods to be defined on `UploadedFile` (@janko)